updated instruments routes with instrumentEditFunctionality
