+++ routes/instrument.js
@@ -1,47 +1,14 @@
 const utils = require('../utils.js');
+const controller = require('../controllers/instrumentEditsFunctionality.js');
 const express = require('express');
 const mysql = require('../dbcon.js');
 const router = express.Router();
 
 
 // updates an instrument and associated level and returns true if the update was successful
-router.post('/update',utils.checkAuthenticated,(req, res, next) => {
-    try {
-        mysql.pool.query(
-            'UPDATE ProfileInstruments SET LevelID = ?, LastUpdated = NOW() WHERE ProfileID = ? AND InstrumentID = ?',
-            [req.body.levelId, req.body.ProfileKey, req.body.instrumentId],
-            function(err, result) {
-                if(err) {
-                    throw (err);
-                } else if (result.affectedRows === 1) {
-                    res.send(true);
-                } else {
-                    throw(new ReferenceError("No such instrument for this user"));
-                }
-        });
-    } catch(err) {
-        res.redirect(utils.profileUpdateErrorRedirect());
-    }
-});
+router.post('/update',utils.checkAuthenticated, controller.updateInstrumentsOnProfile);
 
 // inserts an instrument and associated level and returns true if the insert was successful
-router.post('/add', utils.checkAuthenticated, (req, res, next) => {
-    try {
-        mysql.pool.query(
-            'INSERT INTO ProfileInstruments (ProfileID, InstrumentID, LevelID, CreateDate) VALUES (?, ?, ?, NOW())',
-            [req.body.ProfileKey, req.body.instrumentId, req.body.levelId],
-            function (err, result) {
-                if (err) {
-                    throw (err);
-                } else if (result.affectedRows === 1) {
-                    res.send(true);
-                } else {
-                    throw (new ReferenceError("Must save profile before adding instruments."));
-                }
-            });
-    } catch (err) {
-        res.redirect(utils.profileUpdateErrorRedirect());
-    }
-});
+router.post('/add', utils.checkAuthenticated, controller.addInstrumentsOnProfile);
 
 module.exports = router
